fix: feedbacks + add drizzle studio
