+++ README.md
@@ -1,1 +1,7 @@
 # Hackathon 2024
+
+## Drizzle studio
+
+If you don't want to use phpmyadmin, you can use Drizzle Studio to manage the database. It's a web-based database management tool that allows you to manage your database using a web interface.
+
+You juste need to go to [https://local.drizzle.studio](https://local.drizzle.studio) to access it.
