+++ compose.yml
@@ -15,6 +15,7 @@ services:
     ports:
       - 5173:5173
       - 4173:4173
+      - 4983:4983
 
   mysql:
     image: mysql:8.4.0
