+++ src/lib/server/auth.ts
@@ -18,16 +18,3 @@ export const lucia = new Lucia(adapter, {
     };
   },
 });
-
-declare module 'lucia' {
-  interface Register {
-    Lucia: typeof lucia;
-    DatabaseUserAttributes: DatabaseUserAttributes;
-  }
-}
-
-interface DatabaseUserAttributes {
-  email: string;
-  firstName: string;
-  lastName: string;
-}
