+++ src/lib/server/db/schema/extend-schema/identifiable.ts
@@ -1,7 +1,7 @@
 import { randomUUID } from 'crypto';
 import { varchar } from 'drizzle-orm/mysql-core';
 
-export const idable = {
+export const identifiable = {
   id: varchar('id', { length: 255 })
     .$defaultFn(() => randomUUID())
     .primaryKey(),
