+++ src/lib/server/db/schema/messages.ts
@@ -1,15 +1,15 @@
 import { relations } from 'drizzle-orm';
-import { mysqlTable, text, varchar } from 'drizzle-orm/mysql-core';
+import { json, mysqlTable, varchar } from 'drizzle-orm/mysql-core';
 
-import { idable } from './extend-schema/idable';
+import { identifiable } from './extend-schema/identifiable';
 import { timestampable } from './extend-schema/timestampable';
 import { patients } from './patients';
 import { users } from './users';
 
 export const messages = mysqlTable('messages', {
-  ...idable,
-  content: text('content').notNull(),
-  data: text('data').notNull(),
+  ...identifiable,
+  content: json('content').notNull(),
+  data: json('data').notNull(),
   userId: varchar('user_id', { length: 255 }).references(() => users.id),
   patientId: varchar('patient_id', { length: 255 }).references(() => patients.id),
   ...timestampable,
