+++ src/lib/server/db/schema/patients.ts
@@ -1,13 +1,13 @@
 import { relations } from 'drizzle-orm';
 import { mysqlTable, varchar } from 'drizzle-orm/mysql-core';
 
-import { idable } from './extend-schema/idable';
+import { identifiable } from './extend-schema/identifiable';
 import { timestampable } from './extend-schema/timestampable';
 import { messages } from './messages';
 import { responses } from './responses';
 
 export const patients = mysqlTable('patients', {
-  ...idable,
+  ...identifiable,
   phone: varchar('phone', { length: 10 }).unique().notNull(),
   ...timestampable,
 });
